Can now resize storage

+++ Point.h
@@ -2,11 +2,11 @@
 class Point
 {
 public:
-	Point(float first = -1, float second = -1)
+	Point(double first = -1, double second = -1)
 	{
 		x = first;
 		y = second;
 	}
-	float x;
-	float y;
+	double x;
+	double y;
 };
